Refactor token APIFields models
