+++ src/shared/models/apiFields/getToken.ts
@@ -1,0 +1,5 @@
+import { SHARED_APIFields__BASE } from '.'
+
+export interface SHARED_APIFields__GetToken extends SHARED_APIFields__BASE {
+    accessToken?: string
+}
