+++ src/shared/models/apiFields/validateToken.ts
@@ -1,0 +1,3 @@
+import { SHARED_APIFields__BASE } from '.'
+
+export type SHARED_APIFields__ValidateToken = SHARED_APIFields__BASE
